Update wordpress configuration with new source folders and task options
The commit adds new source folders to the docker-tfs.iml file for the Action_Scheduler and PHPMailer package prefixes. It also introduces a new task option to the watcherTasks.xml file for handling SCSS files in a new Wordpress theme. Lastly, it adds a new directory mapping to the vcs.xml file for managing VCS with the new Wordpress theme.